requirements.txt
android==1.0
asgiref==3.4.1
astroid==2.4.1
audiostream==0.2
certifi==2021.5.30
charset-normalizer==2.0.3
cmake==3.17.2
cycler==0.10.0
defusedxml==0.7.1
Django==3.2.5
djangorestframework==3.12.4
httpie==2.5.0
idna==3.2
isort==4.3.21
jedi==0.17.0
joblib==0.14.1
Kivy==1.11.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.4
pandas==1.0.3
parso==0.7.0
Pillow==8.3.2
pygame==2.0.0.dev8
pygame-sdl2==2.1.0
Pygments==2.10.0
pyjnius==1.2.1
pylint==2.5.2
pyparsing==2.4.7
PySDL2==0.9.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.9.1
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.26.6
wrapt==1.12.1
